+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/callable/function/Function.java
@@ -24,9 +24,11 @@ import org.enso.interpreter.node.callable.dispatch.InvokeFunctionNode;
 import org.enso.interpreter.node.expression.builtin.BuiltinRootNode;
 import org.enso.interpreter.runtime.Context;
 import org.enso.interpreter.runtime.callable.CallerInfo;
+import org.enso.interpreter.runtime.callable.UnresolvedConversion;
 import org.enso.interpreter.runtime.callable.UnresolvedSymbol;
 import org.enso.interpreter.runtime.callable.argument.ArgumentDefinition;
 import org.enso.interpreter.runtime.callable.argument.Thunk;
+import org.enso.interpreter.runtime.callable.atom.AtomConstructor;
 import org.enso.interpreter.runtime.library.dispatch.MethodDispatchLibrary;
 import org.enso.interpreter.runtime.state.data.EmptyMap;
 import org.enso.interpreter.runtime.data.Array;
@@ -401,6 +403,55 @@ public final class Function implements TruffleObject {
       Function function = doResolve(context, symbol);
       if (function == null) {
         throw new MethodDispatchLibrary.NoSuchMethodException();
+      }
+      return function;
+    }
+  }
+
+  @ExportMessage
+  boolean canConvertFrom() {
+    return true;
+  }
+
+  @ExportMessage
+  static class GetConversionFunction {
+
+    static final int CACHE_SIZE = 10;
+
+    @CompilerDirectives.TruffleBoundary
+    static Function doResolve(Context context, AtomConstructor target, UnresolvedConversion conversion) {
+      return conversion.resolveFor(target, context.getBuiltins().function(), context.getBuiltins().any());
+    }
+
+    @Specialization(
+            guards = {
+                    "!context.isInlineCachingDisabled()",
+                    "cachedTarget == target",
+                    "cachedConversion == conversion",
+                    "function != null"
+            },
+            limit = "CACHE_SIZE")
+    static Function resolveCached(
+            Function _this,
+            AtomConstructor target,
+            UnresolvedConversion conversion,
+            @CachedContext(Language.class) Context context,
+            @Cached("conversion") UnresolvedConversion cachedConversion,
+            @Cached("target") AtomConstructor cachedTarget,
+            @Cached("doResolve(context, cachedTarget, cachedConversion)") Function function) {
+      return function;
+    }
+
+    @Specialization(replaces = "resolveCached")
+    static Function resolve(
+            Function _this,
+            AtomConstructor target,
+            UnresolvedConversion conversion,
+            @CachedContext(Language.class) Context context)
+            throws MethodDispatchLibrary.NoSuchConversionException {
+      Function function = doResolve(context, target, conversion);
+      if (function == null) {
+        throw new MethodDispatchLibrary.NoSuchConversionException();
       }
       return function;
     }