+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/data/Array.java
@@ -11,7 +11,9 @@ import com.oracle.truffle.api.library.ExportLibrary;
 import com.oracle.truffle.api.library.ExportMessage;
 import org.enso.interpreter.Language;
 import org.enso.interpreter.runtime.Context;
+import org.enso.interpreter.runtime.callable.UnresolvedConversion;
 import org.enso.interpreter.runtime.callable.UnresolvedSymbol;
+import org.enso.interpreter.runtime.callable.atom.AtomConstructor;
 import org.enso.interpreter.runtime.callable.function.Function;
 import org.enso.interpreter.runtime.library.dispatch.MethodDispatchLibrary;
 
@@ -156,6 +158,56 @@ public class Array implements TruffleObject {
       Function function = doResolve(context, symbol);
       if (function == null) {
         throw new MethodDispatchLibrary.NoSuchMethodException();
+      }
+      return function;
+    }
+  }
+
+  @ExportMessage
+  static boolean canConvertFrom(Array receiver) {
+    return true;
+  }
+
+  @ExportMessage
+  static class GetConversionFunction {
+    static final int CACHE_SIZE = 10;
+
+    @CompilerDirectives.TruffleBoundary
+    static Function doResolve(
+        Context context, AtomConstructor target, UnresolvedConversion conversion) {
+      return conversion.resolveFor(
+          target, context.getBuiltins().mutable().array(), context.getBuiltins().any());
+    }
+
+    @Specialization(
+        guards = {
+          "!context.isInlineCachingDisabled()",
+          "cachedConversion == conversion",
+          "cachedTarget == target",
+          "function != null"
+        },
+        limit = "CACHE_SIZE")
+    static Function resolveCached(
+        Array _this,
+        AtomConstructor target,
+        UnresolvedConversion conversion,
+        @CachedContext(Language.class) Context context,
+        @Cached("conversion") UnresolvedConversion cachedConversion,
+        @Cached("target") AtomConstructor cachedTarget,
+        @Cached("doResolve(context, cachedTarget, cachedConversion)") Function function) {
+      return function;
+    }
+
+    @Specialization(replaces = "resolveCached")
+    static Function resolve(
+        Array _this,
+        AtomConstructor target,
+        UnresolvedConversion conversion,
+        @CachedContext(Language.class) Context context)
+        throws MethodDispatchLibrary.NoSuchConversionException {
+      Function function = doResolve(context, target, conversion);
+      if (function == null) {
+        throw new MethodDispatchLibrary.NoSuchConversionException();
       }
       return function;
     }