+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/library/dispatch/DefaultDoubleExports.java
@@ -6,10 +6,13 @@ import com.oracle.truffle.api.dsl.CachedContext;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.library.ExportLibrary;
 import com.oracle.truffle.api.library.ExportMessage;
+import com.oracle.truffle.api.library.GenerateLibrary;
 import org.enso.interpreter.Language;
 import org.enso.interpreter.runtime.Context;
 import org.enso.interpreter.runtime.builtin.Number;
+import org.enso.interpreter.runtime.callable.UnresolvedConversion;
 import org.enso.interpreter.runtime.callable.UnresolvedSymbol;
+import org.enso.interpreter.runtime.callable.atom.AtomConstructor;
 import org.enso.interpreter.runtime.callable.function.Function;
 
 @ExportLibrary(value = MethodDispatchLibrary.class, receiverType = Double.class)
@@ -53,6 +56,62 @@ public class DefaultDoubleExports {
       Function function = doResolve(context, symbol);
       if (function == null) {
         throw new MethodDispatchLibrary.NoSuchMethodException();
+      }
+      return function;
+    }
+  }
+
+  @ExportMessage
+  public static boolean canConvertFrom(Double receiver) {
+    return true;
+  }
+
+  @ExportMessage
+  public static boolean hasSpecialConversion(Double receiver) {
+    return false;
+  }
+
+  @ExportMessage
+  static class GetConversionFunction {
+    @CompilerDirectives.TruffleBoundary
+    static Function doResolve(
+        Context context, AtomConstructor target, UnresolvedConversion conversion) {
+      Number number = context.getBuiltins().number();
+      return conversion.resolveFor(
+          target, number.getDecimal(), number.getNumber(), context.getBuiltins().any());
+    }
+
+    static final int CACHE_SIZE = 10;
+
+    @Specialization(
+        guards = {
+          "!context.isInlineCachingDisabled()",
+          "cachedConversion == conversion",
+          "cachedTarget == target",
+          "function != null"
+        },
+        limit = "CACHE_SIZE")
+    static Function resolveCached(
+        Double _this,
+        AtomConstructor target,
+        UnresolvedConversion conversion,
+        @CachedContext(Language.class) Context context,
+        @Cached("conversion") UnresolvedConversion cachedConversion,
+        @Cached("target") AtomConstructor cachedTarget,
+        @Cached("doResolve(context, cachedTarget, cachedConversion)") Function function) {
+      return function;
+    }
+
+    @Specialization(replaces = "resolveCached")
+    static Function resolve(
+        Double _this,
+        AtomConstructor target,
+        UnresolvedConversion conversion,
+        @CachedContext(Language.class) Context context)
+        throws MethodDispatchLibrary.NoSuchConversionException {
+      Function function = doResolve(context, target, conversion);
+      if (function == null) {
+        throw new MethodDispatchLibrary.NoSuchConversionException();
       }
       return function;
     }