+++ RELEASES.md
@@ -10,6 +10,12 @@
 - Updated the Scala compiler and dependencies
   ([#3214](https://github.com/enso-org/enso/pull/3214)).
 
+## Interpreter/Runtime
+
+- Added support for overloaded conversions. This allows the `from` method to be
+  implemented with several overloads.
+  ([#3227](https://github.com/enso-org/enso/pull/3227))
+
 ## Enso 0.2.31 (2021-10-01)
 
 ## Interpreter/Runtime