+++ app/gui/CHANGELOG.md
@@ -44,6 +44,12 @@
 [3230]: https://github.com/enso-org/enso/pull/3230
 [3240]: https://github.com/enso-org/enso/pull/3240
 
+#### Enso Compiler
+
+- [Added overloaded `from` conversions.][3227]
+
+[3227]: https://github.com/enso-org/enso/pull/3227
+
 # Enso 2.0.0-alpha.18 (2021-10-12)
 
 <br/>![New Features](/docs/assets/tags/new_features.svg)